refactor(payload): move myExperience from SiteContent to AboutPage global
- Add myExperience richText field to AboutPage global
- Move MY_EXPERIENCE seed data from site-content.seed.ts to page-metadata.seed.ts
- Remove myExperience field and newsletter group from SiteContent global
- Update About page to use page.myExperience instead of siteContent
